# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Shopventory REST API
  description: |-
    REST API for Shopventory, an inventory management web application for small businesses. It exposes endpoints for authentication, products and stock movements, orders, providers, employees, user management and dashboard metrics.

    Authentication is based on JWT. After a successful login the access and refresh tokens are stored in HttpOnly cookies (`AuthToken` and `RefreshToken`); a `Bearer` token in the `Authorization` header is also accepted. Access is role-based (`ADMIN` and `USER`) and list endpoints are paginated.
  contact:
    name: Alberto — Shopventory (TFG)
  version: 0.0.1-SNAPSHOT
servers:
- url: https://localhost:8443
  description: Local development server
security:
- cookieAuth: []
...
paths:
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "403":
          description: Admin role required
        "404":
          description: User not found
    put:
      tags:
      - Users
      summary: Update user
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationDTO"
        required: true
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Invalid user data
        "403":
          description: Admin role required
        "404":
          description: User not found
    delete:
      tags:
      - Users
      summary: Delete user
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: User deleted
        "403":
          description: Admin role required
        "404":
          description: User not found
...
  /api/users/{id}/enabled:
    post:
      tags:
      - Users
      summary: Enable or disable a user account
      operationId: setEnabled
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: value
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Account status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "403":
          description: Admin role required
        "404":
          description: User not found
  /api/users/{id}/approve:
    post:
      tags:
      - Users
      summary: Approve pending user registration
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "403":
          description: Admin role required
        "404":
          description: User not found
...
  /api/users:
    get:
      tags:
      - Users
      summary: List all users paginated
      operationId: list_4
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: User list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponse"
        "403":
          description: Admin role required
  /api/users/pending:
    get:
      tags:
      - Users
      summary: Get list of users pending approval
      operationId: pending
      responses:
        "200":
          description: Pending users list returned
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Admin role required
...
  /api/dashboard/stock-movements:
    get:
      tags:
      - Dashboard
      summary: Get all stock movements paginated
      operationId: stockMovements
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Stock movement list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponse"
        "401":
          description: Not authenticated
  /api/dashboard/stats:
    get:
      tags:
      - Dashboard
      summary: Get dashboard statistics summary
      operationId: stats
      responses:
        "200":
          description: Dashboard statistics returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStatsDTO"
        "401":
          description: Not authenticated
...
components:
  schemas:
    UserRegistrationDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        name:
          type: string
        surname:
          type: string
        phone:
          type: string
      required:
      - email
      - name
      - password
      - surname
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: string
        approved:
          type: boolean
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    DashboardStatsDTO:
      type: object
      properties:
        totalProducts:
          type: integer
          format: int64
        lowStockProducts:
          type: integer
          format: int64
        totalProviders:
          type: integer
          format: int64
        totalUsers:
          type: integer
          format: int64
        pendingApprovals:
          type: integer
          format: int64
        totalOrders:
          type: integer
          format: int64
        pendingSales:
          type: integer
          format: int64
        pendingPurchases:
          type: integer
          format: int64
        totalSalesAmount:
          type: number
        topProducts:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        lowestStockProducts:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        categoryDistribution:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: "JWT access token stored in an HttpOnly cookie, issued by POST\
        \ /api/auth/login."
      name: AuthToken
      in: cookie
    bearerAuth:
      type: http
      description: JWT access token sent in the Authorization header as 'Bearer <token>'.
      scheme: bearer
      bearerFormat: JWT